Change media player files name
